services/requirements.txt
# ChromaDB Search Service Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0

# Web API dependencies
flask>=2.3.0
flask-cors>=4.0.0

# Optional: for better performance
numpy>=1.24.0
torch>=2.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
